// src/core/enums/departmentEnums/department.enum.ts
export enum CompanyDepartmentEnum {
    COMPRAS = "Compras",
    VENTAS = "Ventas",
    ALMACEN = "Almacén",
    LOGISTICA = "Logística",
    MARKETING = "Marketing",
    ATENCION_AL_CLIENTE = "Atención al Cliente",
    FINANZAS = "Finanzas",
    ADMINISTRACION = "Administración",
    RECURSOS_HUMANOS = "Recursos Humanos",
    TECNOLOGIA_DE_LA_INFORMACION = "Tecnología de la Información",
    CONTROL_DE_CALIDAD = "Control de Calidad",
    CONTABILIDAD = "Contabilidad", // Podría ser parte de Finanzas, pero a veces es un departamento separado
    LEGAL = "Legal", // Si la empresa tiene un departamento legal
    PLANIFICACION = "Planificación", // Para la planificación de la demanda, etc.
    IMPORTACIONES = "Importaciones", // Si la empresa compra productos del extranjero
    EXPORTACIONES = "Exportaciones", // Si la empresa vende productos al extranjero
    SERVICIO_TECNICO = "Servicio Técnico", // Si los productos requieren soporte técnico
  }
  